Move FieldMeta into datatypes, implement anyhow error rpepr
