+++ marrow/src/error.rs
@@ -1,10 +1,21 @@
 //! Error handling in `marrow`
 //!
 
-/// A result type that defaults to `marrow::Error`
+/// A result type that defaults to [`MarrowError`]
 pub type Result<T, E = MarrowError> = std::result::Result<T, E>;
 
 /// Errors that may occur in `marrow`
+///
+/// The error carries a backtrace if `RUST_BACKTRACE=1`, see [`std::backtrace`] for details. The
+/// backtrace can be retrieved using [`backtrace()`][`MarrowError::backtrace()`]. If the error is
+/// caused by another error, that error can be retrieved with
+/// [`source()`][std::error::Error::source].
+///
+/// # Display representation
+///
+/// This error type follows anyhow's display representation: when printed with display format (`{}`)
+/// (or converted to string) the error does not include a backtrace. Use the debug format (`{:?}`)
+/// to include the backtrace information.
 pub struct MarrowError(Box<ErrorImpl>);
 
 /// The kind of error to simplify matching against known error conditions
@@ -62,6 +73,13 @@ impl std::error::Error for MarrowError {
             None => None,
         }
     }
+
+    fn cause(&self) -> Option<&dyn std::error::Error> {
+        match &self.0.cause {
+            Some(error) => Some(error.as_ref()),
+            None => None,
+        }
+    }
 }
 
 struct ErrorImpl {
@@ -73,7 +91,12 @@ struct ErrorImpl {
 
 impl std::fmt::Debug for MarrowError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Error: {}", self.0.message)
+        write!(
+            f,
+            "Error: {message}\n{backtrace}",
+            message = self.0.message,
+            backtrace = BacktraceDisplay(&self.0.backtrace),
+        )
     }
 }
 
@@ -83,13 +106,25 @@ impl std::fmt::Display for MarrowError {
     }
 }
 
+struct BacktraceDisplay<'a>(&'a Backtrace);
+
+impl<'a> std::fmt::Display for BacktraceDisplay<'a> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self.0.status() {
+            BacktraceStatus::Captured => write!(f, "Backtrace:\n{bt}", bt=self.0),
+            BacktraceStatus::Disabled => write!(f, "Backtrace not captured; set the `RUST_BACKTRACE=1` env variable to enable"),
+            _ => write!(f, "Backtrace not captured: most likely backtraces are not supported on the current platform"),
+        }
+    }
+}
+
 macro_rules! fail {
     ($kind:expr, $($msg:tt)*) => {
         return Err($crate::error::error_with_kind_and_message($kind, format!($($msg)*)))
     };
 }
 
-use std::backtrace::Backtrace;
+use std::backtrace::{Backtrace, BacktraceStatus};
 
 pub(crate) use fail;
